+++ CHANGELOG.md
@@ -8,11 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - reply-to address to emails, using tenant contact email address (system settings)
 - buttons to reset forms
+- validation for tenant database urls, when creating or updating a tenant
 
 ### Changed
 
 - when signing up for a session with follow-ups, the contact receives one confirmation message containing all sessions
 - using uuids instead of ids as foreign keys in assignment table
+- tenant description, icon and styles are optional
 
 ### Fixed